requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0
selenium>=4.0.0
undetected-chromedriver>=3.5.0
fake-useragent>=1.4.0
beautifulsoup4>=4.11.0
matplotlib>=3.5.0
seaborn>=0.11.0
ipython>=8.0.0
jupyter>=1.0.0
streamlit>=1.28.0
plotly>=5.15.0
openpyxl>=3.1.0
agno>=0.1.0
python-dotenv>=1.0.0
python-docx>=0.8.11
reportlab>=4.0.0
